todo_drf/api/requirements_ENT.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
attrs=19.3.0=py_0
backcall=0.1.0=py36_0
beautifulsoup4=4.9.0=pypi_0
bleach=3.1.4=py_0
brotlipy=0.7.0=py36h8c4c3a4_1000
ca-certificates=2020.1.1=0
catalogue=1.0.0=py36h9f0ad1d_1
certifi=2020.4.5.1=pypi_0
cffi=1.11.5=py36_0
chardet=3.0.4=py36h9f0ad1d_1006
click=7.1.2=pypi_0
cryptography=2.9.2=py36h45558ae_0
cycler=0.10.0=pypi_0
cymem=2.0.3=py36h831f99a_1
cython-blis=0.4.1=py36h8c4c3a4_1
dataclasses=0.7=pypi_0
dbus=1.13.12=h746ee38_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
filelock=3.0.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
glib=2.56.2=hd408876_0
gmp=6.1.2=hb3b607b_0
google=2.0.3=pypi_0
google-search=1.0.2=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=pypi_0
icu=58.2=he6710b0_3
idna=2.9=py_1
importlib_metadata=1.5.0=py36_0
ipykernel=5.1.4=py36h39e3cac_0
ipython=7.13.0=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.5.1=py_0
jedi=0.17.0=py36_0
jinja2=2.11.2=py_0
joblib=0.14.1=pypi_0
jpeg=9b=habf39ab_1
jsonschema=3.2.0=py36_0
jupyter=1.0.0=py36_7
jupyter_client=6.1.3=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.3=py36_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.1=pypi_0
kiwisolver=1.2.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_5
liblapack=3.8.0=14_openblas
libopenblas=0.3.7=h5ec1e0e_6
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
lxml=4.5.0=pypi_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.2.1=pypi_0
mistune=0.8.4=py36h7b6447c_0
murmurhash=1.0.0=py36h831f99a_0
nbconvert=5.6.1=py36_0
nbformat=5.0.6=py_0
ncurses=6.2=he6710b0_1
nltk=3.5=pypi_0
notebook=6.0.3=py36_0
numpy=1.18.4=py36h7314795_0
openssl=1.1.1g=h7b6447c_0
pandas=1.0.3=py36h0573a6f_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.7.0=py_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_0
pip=20.0.2=py36_3
plac=0.9.6=py36_0
preshed=3.0.2=py36h831f99a_2
prometheus_client=0.7.1=py_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
protobuf=3.11.3=pypi_0
ptyprocess=0.6.0=py36_0
pycparser=2.20=py_0
pygments=2.6.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pypi_0
pyqt=5.9.2=py36h22d08a2_1
pyrsistent=0.16.0=py36h7b6447c_0
pysocks=1.7.1=py36h9f0ad1d_1
python=3.6.10=h7579374_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytz=2020.1=py_0
pyyaml=5.3.1=pypi_0
pyzmq=18.1.1=py36he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.7.3=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
regex=2020.5.7=pypi_0
requests=2.23.0=pyh8c360ce_2
sacremoses=0.0.43=pypi_0
scikit-learn=0.23.0=pypi_0
scipy=1.4.1=pypi_0
send2trash=1.5.0=py36_0
sentence-transformers=0.2.6.1=pypi_0
sentencepiece=0.1.90=pypi_0
seqeval=0.0.12=pypi_0
setuptools=46.2.0=py36_0
simpletransformers=0.10.4=pypi_0
sip=4.19.13=py36he6710b0_0
six=1.14.0=py36_0
soupsieve=2.0=pypi_0
spacy=2.2.3=py36hc9558a2_0
sqlite=3.31.1=h62c20be_1
srsly=1.0.2=py36h831f99a_0
tabulate=0.8.7=pypi_0
tensorboardx=2.0=pypi_0
terminado=0.8.3=py36_0
testpath=0.4.4=py_0
thinc=7.3.0=py36hc9558a2_0
threadpoolctl=2.0.0=pypi_0
tk=8.6.8=hbc83047_0
tokenizers=0.7.0=pypi_0
torch=1.5.0=pypi_0
tornado=6.0.4=py36h7b6447c_1
tqdm=4.46.0=pyh9f0ad1d_0
traitlets=4.3.3=py36_0
transformers=2.9.0=pypi_0
tweet-preprocessor=0.6.0=pypi_0
urllib3=1.25.9=py_0
wasabi=0.6.0=py_0
wcwidth=0.1.9=py_0
webencodings=0.5.1=py36_1
wheel=0.34.2=py36_0
widgetsnbextension=3.5.1=py36_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.1=he6710b0_3
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
